requirements.txt
# Live Caption - Requirements
# Python 3.11+

# GUI Framework
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Audio Processing
PyAudio==0.2.14
numpy==1.26.4
scipy==1.12.0

# Speech Recognition (STT)
faster-whisper==1.0.3
torch==2.2.0
torchaudio==2.2.0

# Translation
transformers==4.38.2
sentencepiece==0.2.0
sacremoses==0.1.1

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.1

# Utilities
loguru==0.7.2
psutil==5.9.8
requests==2.31.0

# Development & Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-qt==4.4.0
black==24.2.0
flake8==7.0.0

# Packaging
pyinstaller==6.4.0
